requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
httpx==0.27.0
qdrant-client==1.9.0
asyncpg==0.29.0
python-multipart==0.0.9
pydantic-settings==2.3.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
